+++ tests/testthat/resources/magrittr.R
@@ -1,0 +1,4 @@
+'A' %>% library
+'B' %>% library()
+'C' %>% library(.)
+