+++ tests/testthat/test-dependencies.R
@@ -327,3 +327,8 @@ test_that("renv.ignore=FALSE, eval=TRUE is handled", {
   expect_true("A" %in% deps$Package)
   expect_false("a" %in% deps$Package)
 })
+
+test_that("piped expressions can be parsed for dependencies", {
+  deps <- dependencies("resources/magrittr.R", progress = FALSE)
+  expect_setequal(deps$Package, c("A", "B", "C"))
+})